revert moving solution
